fix: renaming and refactoring project architecture
