+++ quantum-core/src/main/java/control/command/circuitCommand/exportCommand/ExportJSON.java
@@ -1,0 +1,25 @@
+package control.command.circuitCommand.exportCommand;
+
+import control.command.QuantumCommand;
+
+public class ExportJSON implements QuantumCommand {
+    @Override
+    public void execute() {
+
+    }
+
+    @Override
+    public void undo() {
+
+    }
+
+    @Override
+    public boolean canUndo() {
+        return false;
+    }
+
+    @Override
+    public void redo() {
+
+    }
+}
