+++ quantum-core/src/main/java/control/command/circuitCommand/importCommand/ImportQiskit.java
@@ -1,0 +1,25 @@
+package control.command.circuitCommand.importCommand;
+
+import control.command.QuantumCommand;
+
+public class ImportQiskit implements QuantumCommand {
+    @Override
+    public void execute() {
+
+    }
+
+    @Override
+    public void undo() {
+
+    }
+
+    @Override
+    public boolean canUndo() {
+        return false;
+    }
+
+    @Override
+    public void redo() {
+
+    }
+}
